fix: content dto/interfaces
