+++ project/libs/shared/core/src/lib/types/blog/posts/content/base-post-content.interface.ts
@@ -1,0 +1,4 @@
+export interface BasePostContent {
+  id?: string;
+  createdAt?: Date;
+}
