--- project/libs/shared/core/src/lib/types/blog/posts/content/index.ts
@@ -1,48 +1,0 @@
-import { PartialType, getSchemaPath } from '@nestjs/swagger';
-import { LinkPostContent } from './link-post-content.interface';
-import { PhotoPostContent } from './photo-post-content.interface';
-import { QuotePostContent } from './quote-post-content.interface';
-import { TextPostContent } from './text-post-content.interface';
-import { VideoPostContent } from './video-post-content.interface';
-import { PostType } from '../post-type.enum';
-
-export const AllPostContentArray = [
-  VideoPostContent,
-  PhotoPostContent,
-  QuotePostContent,
-  TextPostContent,
-  LinkPostContent,
-];
-
-export type PostContent =
-  | VideoPostContent
-  | LinkPostContent
-  | PhotoPostContent
-  | QuotePostContent
-  | TextPostContent;
-
-export const RefPostContentArray = AllPostContentArray.map((PostTypeClass) => ({
-  $ref: getSchemaPath(PostTypeClass),
-}));
-
-export const AllOptionPostContentArray = [
-  class LinkPostContentOptional extends PartialType(LinkPostContent) {},
-  class PhotoPostContentOptional extends PartialType(PhotoPostContent) {},
-  class VideoPostContentOptional extends PartialType(VideoPostContent) {},
-  class TextPostContentOptional extends PartialType(TextPostContent) {},
-  class QuotePostContentOptional extends PartialType(QuotePostContent) {},
-];
-
-export const RefOptionalPostContentArray = AllOptionPostContentArray.map(
-  (PostTypeClass) => ({
-    $ref: getSchemaPath(PostTypeClass),
-  })
-);
-
-export const PostTypeContent = {
-  [PostType.VIDEO]: VideoPostContent,
-  [PostType.TEXT]: TextPostContent,
-  [PostType.QUOTE]: QuotePostContent,
-  [PostType.LINK]: LinkPostContent,
-  [PostType.PHOTO]: PhotoPostContent,
-};
