+++ project/libs/shared/core/src/lib/types/blog/posts/content/photo-post-content.interface.ts
@@ -1,14 +1,5 @@
-import { Expose } from 'class-transformer';
-import { BasePostContent } from './post-content.interface';
-import { ApiProperty } from '@nestjs/swagger';
-import { IsString } from 'class-validator';
+import { BasePostContent } from './base-post-content.interface';
 
-export class PhotoPostContent extends BasePostContent {
-  @Expose()
-  @ApiProperty({
-    description: 'Photo image ID',
-    example: '65b809b8d6443b043b33eedb',
-  })
-  @IsString()
-  imageId!: string;
+export interface PhotoPostContent extends BasePostContent {
+  imageId: string;
 }
