--- project/libs/shared/core/src/lib/types/blog/posts/content/post-content.interface.ts
@@ -1,9 +1,0 @@
-import { Expose } from 'class-transformer';
-
-export class BasePostContent {
-  @Expose()
-  id?: string;
-
-  @Expose()
-  createdAt?: Date;
-}
