+++ project/libs/shared/core/src/lib/types/blog/posts/content/video-post-content.interface.ts
@@ -1,25 +1,7 @@
-import { Expose } from 'class-transformer';
-import { BasePostContent } from './post-content.interface';
-import { ApiProperty } from '@nestjs/swagger';
-import { IsString, Matches, MaxLength, MinLength } from 'class-validator';
-import { PostContentValidator } from '@project/validation';
+import { BasePostContent } from './base-post-content.interface';
 
-export class VideoPostContent extends BasePostContent {
-  @ApiProperty({
-    description: 'Video title',
-    example: 'Lorem ipsum dolor sit amet. Suscipit, est?',
-  })
-  @Expose()
-  @IsString()
-  @MinLength(PostContentValidator.video.title.Min)
-  @MaxLength(PostContentValidator.video.title.Max)
-  title!: string;
+export interface VideoPostContent extends BasePostContent {
+  title: string;
 
-  @ApiProperty({
-    description: 'Valid url link',
-    example: 'https://www.youtube.com/watch?v=JU79n3yU1aA',
-  })
-  @Expose()
-  @Matches(PostContentValidator.video.url.Matches)
-  url!: string;
+  url: string;
 }
